Rewrite rels to be one-sided, maybe to-many
Change relationships to have a source and a target rather than a generic
"left" and "right" side. This allows for relationships to be simpler,
removes the concept of a Binding (which is basically just what
relationships are now).

We also add a concept of a "link" between two record that can be refined
in future work for performance. Basically we treat all records as going
across a (doubly indexed) join table.

This is suboptimal performance for 1-N and 1-1 relationships, but vastly
simplifying for now. And there are things we can do later to claw back
performance (like using pointers between records)

The advantage for now is that we can support many-to-many relationships.

Many to-dos are left at this point, but we (hopefully) have a simpler
base to build upon after this change.