Rewrite the way db primitives are managed
This is a pretty significant change to the DB package. Rather than have
db internals work via translating records to go structs, we introduce a
separate implementation of each: one version in "go" land (mainly for
bootstrapping) and one version in "db" land that reads values off of
structs and avoids doing a bunch of deserialization.

There were a few motivations for the change:
* It was tough to decide on "how much" to deserialize, leading to a lot
  of extra data lookups for core operations -- in the new model, it's
  just done lazily
* Further, keeping things "lazy" should allow us to keep the models up
  to date with changes mid-transaction
* Further, though this is speculative, we should do a bit less
  allocation this way in core operations

As a nice side effect, the golang literals are a bit more readable now
that they aren't serving two separate roles (being declared as literals
and being used in core DB operations).

I also added a concept of an interfaces to allow for differing
implementations of core primitives (or user-defined concepts). So a
datatype requires a "Function", which may be a native (ie golang)
function, or a starlark function.

One alternative implementation I considered and prototyped involved
and implementing Relationship/Attribute on top of interfaces.
This would be an alternative to the "loaders" you see in this change.

While that approach offers an appealing purity, in pratice I found the
additional complexity pretty significant, and suspect the performance
would be pretty poor, too. I think I could end up revisiting this down
the line if there's a motivation.