Rewrite the hold
Rewrite the hold to better abstract the indexes. Additionally add two
indexes to track model -> interface and model -> relationship edges.
These are denormalized from the links table in order to provide a
separate view for migrations.

In other words, we let the user update the model -> interface link
in the links index, and then when the automigrate handler runs, we
update the hold's view of the world.

The impact is that the migrate handler can have all of the information
it needs available when it runs. A secondary impact is that the Hold is
now able to handle cascades, so this change also cleans up the code for
tx.Delete.