Rewrite the query engine
The new query engine follows database tradition and uses a planned tree
of operation nodes to sequence its work.

The advantages for now are mainly in removing some of the most confusing
and hard to work with code in aft and replacing it with something far
more debuggable, simple and clear.

There may be some modest performance changes; the new system does far
more allocations per query, but might choose slightly more efficient
plans.

In any case, we add some new features to the query engine with the
rewrite: sorting, limiting, and offsetting. And we have a bit more of a
stable base upon which do write a proper optimizing query planner.